Move action client code to tuck_arms.py
