+++ pr2_tuck_arms_action/scripts/tuck_arms.py
@@ -33,5 +33,75 @@
 #
 # Author: Wim Meeussen
 
-from pr2_tuck_arms_action import tuck_arms_main
-tuck_arms_main.main()
+"""
+usage: tuck_arms.py [-l ACTION] [-r ACTION] [-q]
+Options:
+  -l or --left   Action for left arm
+  -r or --right  Action for right arm
+  -q or --quit   Shut down the action client after completing action
+Actions:
+  t or tuck
+  u or untuck
+
+"""
+
+from pr2_common_action_msgs.msg import TuckArmsAction, TuckArmsGoal
+import actionlib
+import getopt
+import rospy
+
+
+def usage():           
+    print __doc__ % vars()               
+    rospy.signal_shutdown("Help requested")
+
+
+def main():
+    action_name = 'tuck_arms'
+    quit_when_finished = False
+  
+    # check for command line arguments, and send goal to action server if required
+    myargs = rospy.myargv()[1:]
+    if len(myargs):
+        goal = TuckArmsGoal()
+        goal.tuck_left = True
+        goal.tuck_right = True
+        opts, args = getopt.getopt(myargs, 'hql:r:', ['quit','left','right'])
+        for arm, action in opts:
+            if arm in ('-l', '--left'):
+                if action in ('tuck', 't'):
+                  goal.tuck_left = True
+                elif action in ('untuck', 'u'):
+                  goal.tuck_left = False
+                else:
+                   rospy.logerr('Invalid action for right arm: %s'%action)
+                   rospy.signal_shutdown("ERROR")
+  
+            if arm in ('-r', '--right'):
+                if action in ('tuck', 't'):
+                    goal.tuck_right = True
+                elif action in ('untuck', 'u'):
+                    goal.tuck_right = False
+                else:
+                    rospy.logerr('Invalid action for left arm: %s'%action)
+                    rospy.signal_shutdown("ERROR")
+  
+            if arm in ('--quit', '-q'):
+                quit_when_finished = True
+  
+            if arm in ('--help', '-h'):
+                usage()
+        
+        tuck_arm_client = actionlib.SimpleActionClient(action_name, TuckArmsAction)
+        rospy.logdebug('Waiting for action server to start')
+        tuck_arm_client.wait_for_server(rospy.Duration())
+        rospy.logdebug('Sending goal to action server')
+        tuck_arm_client.send_goal_and_wait(goal, rospy.Duration(30.0), rospy.Duration(5.0))
+  
+        if quit_when_finished:
+            rospy.signal_shutdown("Quitting")
+
+if __name__ == '__main__':
+    rospy.init_node('tuck_arms_client')
+    main()
+    rospy.spin()
